clientPortalFunctions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issue

azure-functions
psycopg2-binary
logging
jsonschema
